'adminpanel'

+++ app/Http/Controllers/panell/Album/AlbumController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Panell\Album;
 use App\Http\Controllers\Controller;
 use App\Http\Models\Music\Album;
 use App\Http\Models\Music\Artist;
+use App\Http\Models\Music\File;
 use App\Http\Requests\Panell\Album\AlbumStoreRequest;
 use App\Http\Requests\Panell\Album\AlbumUpdateRequest;
 use App\Http\Resources\Panell\Album\AlbumIndexResource;
@@ -21,10 +22,21 @@ class AlbumController extends Controller
 	 */
     public function index()
     {
-        $albums=Album::paginate();
-        return AlbumIndexResource::collection($albums);
+    	if (!auth()->user()->can('view',Album::class)){
+    		abort(403);
+		}
+        $album=Album::paginate()->load('file');
+        $albums=AlbumIndexResource::collection($album);
+        return view('admin.albums.index',compact('albums'));
     }
 
+	public function create()
+	{
+		if (!auth()->user()->can('create',Album::class)){
+			abort(403);
+		}
+		return view('admin.albums.create');
+}
 	/**
 	 * Store a newly created resource in storage.
 	 *
@@ -33,12 +45,29 @@ class AlbumController extends Controller
 	 * @return \Illuminate\Http\Response|array
 	 *
 	 */
+	public function create_song(Album $id)
+	{
+		if (!auth()->user()->can('create',Album::class)){
+			abort(403);
+		}
+
+		$art=Artist::all();
+		return view('admin.albums.create_song',compact('id','art'));
+	}
     public function store(AlbumStoreRequest $request,Album $album)
     {
     	\DB::beginTransaction();
     	try{
 			$album->fill($request->all());
 			$album->save();
+			// ------------------------------------ attach file ------------------------------------
+			$file     = new File();
+			$FileName = $request->file('image')->getClientOriginalName();
+			$request->file('image')->move('images/album_image', $FileName);
+			$file->image = $FileName;
+			$file->save();
+			$album->file()->attach($file->id);
+
 			\DB::commit();
 			return[
 				'success'=>true,
@@ -66,20 +95,51 @@ class AlbumController extends Controller
 	 */
     public function show(Album $album)
     {
-        return new AlbumSowResource($album);
+		if (!auth()->user()->can('view',Album::class)){
+			abort(403);
+		}
+
+		$pure_data = new AlbumShowResource($album->load('file', 'songs'));
+		$album     = json_decode(json_encode($pure_data->resource));
+		if ($album->file) {
+			$path    = explode('/', $album->file[0]->image);
+			$path    = end($path);
+			$storage = '/images/album_image/' . $path;
+		} else {
+			$storage = '/images/song_image/nophoto.png';
+		}
+
+		// ------------------------------------ relation ------------------------------------
+
+
+		$song = $album->songs;
+//		$song=[$song];
+
+//		dd($song);
+		$album = json_decode(json_encode($album), true);
+
+//		dd($album);
+		$album = json_decode(json_encode($album));
+//dd($album);
+		return view('admin.albums.show', compact('song',  'album', 'storage'));
+
 	}
+
 
 	/**
 	 * Update the specified resource in storage.
 	 *
-	 * @param  \Illuminate\Http\Request $request
+	 * @param  \Illuminate\Http\Request $request]
 	 * @param  int                      $id
 
 	 * @return \Illuminate\Http\Response|array
 	 */
     public function update(AlbumUpdateRequest $request, Album $album)
     {
-        \DB::beginTransactin();
+		if (!auth()->user()->can('update',Album::class)){
+			abort(403);
+		}
+        \DB::beginTransaction();
         try{
         	$album->fill($request->all());
         	$album->save();
@@ -106,10 +166,14 @@ class AlbumController extends Controller
 	 */
     public function destroy(Album $album)
     {
-		\DB::beginTransactin();
+		if (!auth()->user()->can('delete',Album::class)){
+			abort(403);
+		}
+		\DB::beginTransaction();
 		try{
 			$album->delete();
-			DB::commit();
+
+			\DB::commit();
 			return[
 				'success'=>true,
 				'message'=>trans('responses.panel.music.message.destroy'),
@@ -125,6 +189,9 @@ class AlbumController extends Controller
     }
 	public function restore($id)
 	{
+		if (!auth()->user()->can('forceDelete',Album::class)){
+			abort(403);
+		}
 		/**
 		 * @mixin Artist
 		 */
@@ -149,6 +216,9 @@ class AlbumController extends Controller
 
 	public function list()
 	{
+		if (!auth()->user()->can('view',Album::class)){
+			abort(403);
+		}
 		return Album::select('id','name')->get();
 	}
 }
